+++ src/main/java/dev/ftb/mods/ftbskies2aerocompanion/mixin/TropicraftRandomThreadSafeMixin.java
@@ -1,8 +1,8 @@
 package dev.ftb.mods.ftbskies2aerocompanion.mixin;
 
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Mutable;
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(ServerLevel.class)
+@Mixin(Level.class)
 public class TropicraftRandomThreadSafeMixin {
 
     private static final ResourceLocation TROPICS =
@@ -22,11 +22,15 @@ public class TropicraftRandomThreadSafeMixin {
     @Mutable
     public RandomSource random;
 
+    @Shadow
+    @Final
+    public boolean isClientSide;
+
     @SuppressWarnings("deprecation")
     @Inject(method = "<init>", at = @At("TAIL"))
     private void ftbskies2aero$threadSafeTropicraftRandom(CallbackInfo ci) {
-        ServerLevel level = (ServerLevel) (Object) this;
-        if (level.dimension().location().equals(TROPICS)) {
+        Level level = (Level) (Object) this;
+        if (!isClientSide && level.dimension().location().equals(TROPICS)) {
             this.random = RandomSource.createThreadSafe();
         }
     }